Added files for Daniel to edit
