+++ src/models/logged-work.js
@@ -1,0 +1,8 @@
+const mongoose = require('mongoose')
+
+const Project = new mongoose.Schema({
+
+})
+
+// export default mongoose.model('user', User)
+module.exports = mongoose.model('loggedWork', Project)
